kg_from_unstructure_text/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
neo4j==5.15.0
nebula3-python==3.5.0
openai==1.10.0
chromadb==0.4.22
sentence-transformers==2.2.2
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
rank-bm25==0.2.2
jieba==0.42.1
python-docx==1.1.0
pypdf==4.0.1
python-pptx==0.6.23
openpyxl==3.1.2
networkx==3.2.1
pytest==8.0.0
httpx==0.26.0
tenacity==8.2.3
tiktoken==0.5.2